pyproject.toml
[tool.poetry]
name = "djangopostgresrl"
version = "0.1.0"
description = ""
authors = ["Sidrah-Madiha <sidrahsms@yahoo.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.2"
psycopg2 = "^2.8.6"
drf-nested-routers = "^0.93.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
